Fixed requestedFor property for auto-scheduled Bitbucket pipelines builds
CI builds generated via Bitbucket pipelines have a "creator" property
when the build is triggered by a commit to the associated Bitbucket
repo. However, for builds that are triggered by a timer, the "creator"
property is absent. Instead, there is a "tigger" property.
This commit fixes the bug that always looked for "creator" property,
even for timer-triggered builds.